Checkpoint (getting list of paired bluetooth devices works)
